// src/types/basket-cameras-data.ts
export type BasketCamerasData = {
    id: number;
    name: string;
    vendorCode: string;
    type: string;
    category: string;
    description: string;
    reviewCount: number;
    level: string;
    price: number;
    previewImg: string;
    previewImg2x: string;
    previewImgWebp: string;
    previewImgWebp2x: string;
    rating: number;
    basketItemCount: number;
}